--- src/ai/backend/manager/sokovan/idle_check/checkers/utilization/__init__.py
@@ -1,1 +1,0 @@
-"""Utilization idle checker."""
